Migrated all code into headers, removed Makevars.
