--- src/Makevars
@@ -1,33 +1,0 @@
-RHDF5LIB_LIBS=`echo 'Rhdf5lib::pkgconfig("PKG_CXX_LIBS")'|\
-	"${R_HOME}/bin/R" --vanilla --slave`
-PKG_LIBS=$(RHDF5LIB_LIBS)
-
-all: $(SHLIB) copying
-
-include Makevars.common
-
-# Wait for R to build the shared object, and then pick up the object files.
-libbeachmat.a: $(SHLIB)
-	ar rc libbeachmat.a $(EXPORT_OBJECTS)
-
-SHLIB_CXXLD=`${R_HOME}/bin/R CMD config SHLIB_CXXLD`
-LDFLAGS=`${R_HOME}/bin/R CMD config LDFLAGS`
-SHLIB_CXXLDFLAGS=`${R_HOME}/bin/R CMD config SHLIB_CXXLDFLAGS` 
-
-libbeachmat.so: $(SHLIB)
-	$(SHLIB_CXXLD) ${SHLIB_CXXLDFLAGS} ${LDFLAGS} -o $@ $(EXPORT_OBJECTS) ${PKG_LIBS} 
- 
-# Specifying copying locations.  
-BEACHMAT_LIBDIR="${R_PACKAGE_DIR}/lib${R_ARCH}"
-BEACHMAT_INCLUDEDIR="${R_PACKAGE_DIR}/include/beachmat"
-
-copying: libbeachmat.a libbeachmat.so
-	mkdir -p $(BEACHMAT_LIBDIR) $(BEACHMAT_INCLUDEDIR)
-	cp $(EXPORT_HEADERS) $(BEACHMAT_INCLUDEDIR)
-	mv libbeachmat.* $(BEACHMAT_LIBDIR)
-
-# Cleaning commands.
-shlib-clean: clean
-
-clean: 
-	rm -f libbeachmat.*
