--- src/Makevars.common
@@ -1,9 +1,0 @@
-# Specifying the headers and objects to put into the exported library.
-EXPORT_HEADERS=dim_checker.h utils.h HDF5_utils.h output_param.h \
-    HDF5_reader.h Csparse_reader.h dense_reader.h simple_reader.h delayed_reader.h unknown_reader.h external_reader.h \
-    simple_writer.h Csparse_writer.h HDF5_writer.h \
-    LIN_matrix.h LIN_methods_read.h LIN_methods_write.h LIN_output.h \
-    logical_matrix.h integer_matrix.h character_matrix.h numeric_matrix.h character_output.h  
-EXPORT_OBJECTS=dim_checker.o character_matrix.o character_output.o integer_matrix.o logical_matrix.o numeric_matrix.o utils.o HDF5_utils.o output_param.o
-
-
