--- src/Makevars.win
@@ -1,35 +1,0 @@
-RHDF5LIB_LIBS=$(shell echo 'Rhdf5lib::pkgconfig("PKG_CXX_LIBS")'|\
-	"${R_HOME}/bin/R" --vanilla --slave)
-PKG_LIBS=$(RHDF5LIB_LIBS)
-PKG_LIBS+=$(shell ${R_HOME}/bin/R CMD config --ldflags)
-
-all: $(SHLIB) copying
-
-include Makevars.common
-
-# Wait for R to build the shared object, and then pick up the object files.
-
-# this is sufficient to build static library with the suitable version of ar
-libbeachmat.a: $(EXPORT_OBJECTS)
-
-SHLIB_CXXLD=$(shell ${R_HOME}/bin/R CMD config SHLIB_CXXLD)
-LDFLAGS=$(shell ${R_HOME}/bin/R CMD config LDFLAGS)
-SHLIB_CXXLDFLAGS=$(shell ${R_HOME}/bin/R CMD config SHLIB_CXXLDFLAGS) 
-
-libbeachmat.dll: $(SHLIB)
-	$(SHLIB_CXXLD) ${SHLIB_CXXLDFLAGS} ${LDFLAGS} -s -static-libgcc -o $@ $(EXPORT_OBJECTS) ${PKG_LIBS} 
- 
-# Specifying copying locations.  
-BEACHMAT_LIBDIR="${R_PACKAGE_DIR}/lib${R_ARCH}"
-BEACHMAT_INCLUDEDIR="${R_PACKAGE_DIR}/include/beachmat"
-
-copying: libbeachmat.dll libbeachmat.a
-	mkdir -p $(BEACHMAT_LIBDIR) $(BEACHMAT_INCLUDEDIR)
-	cp $(EXPORT_HEADERS) $(BEACHMAT_INCLUDEDIR)
-	mv libbeachmat.* $(BEACHMAT_LIBDIR)
-
-# Cleaning commands.
-shlib-clean: clean
-
-clean: 
-	rm -f libbeachmat.*
