--- src/functions.cpp
@@ -1,22 +1,0 @@
-#include "functions.h"
-#include "R_ext/Rdynload.h"
-#include "R_ext/Visibility.h"
-
-#define REGISTER(x, i) {#x, (DL_FUNC) &x, i}
-
-extern "C" {
-
-static const R_CallMethodDef all_call_entries[] = {
-    REGISTER(rechunk_matrix, 8),
-    REGISTER(find_chunks, 1),
-    {NULL, NULL, 0}
-};
-
-void attribute_visible R_init_beachmat(DllInfo *dll) {
-    R_registerRoutines(dll, NULL, all_call_entries, NULL, NULL);
-    R_useDynamicSymbols(dll, FALSE);
-    R_forceSymbols(dll, TRUE);
-}
-
-}
-
